added document automation events
